Note test type (jest or cypress) in slack notify
Uses the `CIRCLE_JOB` built-in envvar to give more specificity to the slack notification. Was the run a jest test or cypress test?

I would still like to do better than this by persisting all junit xml result files to the workspace and then parsing and reporting the total results only once. However, I think we will need a diffeerent strategy to achieve this goal. Circle does let you configure jobs in a DAG, so in theory a parse-and-notify job could be set to run last. Unfortunately, from the docs it looks like Circle requires the previous jobs to complete **successfully** in order for the next job in the DAG to run. This makes sense for most users (don't run the deploy job if the test job failed) but in our use case we want to run parse-and-notify even if (especially if) some of the tests failed.

So for now the solution will have to be more notification spam and regrouping on the bigger goals.